✨feat: JwtAuthenticationFilter authentication check
Simple Filter that should be used before every HTTP request.

Requires HttpServletRequest which is the request already sent by the client where we extract the JWT token.

Requires HttpServletResponse which is the response that will be sent by the server. The filter is called BEFORE the response and changes it depending on the authentication.

Requires a FilterChain which are a sequence of filters that do logical checks on the request and response.

Finally after the filter chain the authentication happens:
It extracts the JWT Token from the Authorization Header if it exists and starts with "Bearer ".

Then it extracts the user's email from the token, if not empty and not already authenticated and finally it compares the user data from the token with the user data from the repository.

The filter extends OncePerRequestFilter to ensure it's only invoked once per request, which is all that's necessary for authentication.